Fix issue # 90: `supervisor file.js` might hang
If the global `node_modules` directory is in a symlinked path
`process.argv[1]` and `__filename` might differ, despite the fact that
they are pointing to the same file.

This commit normalizes `process.argv[1]` to its "realpath" before
checking for its equivalence to `__filename`.

This commit assumes `__filename` has already been normalized to its
realpath by node. If this is not true, `__filename` will need to be
normalized as well.

Note: This has been tested on Windows 7 only.