Added restart-verbose option
Added --restart-verbose / -RV option to log the file that causes the
server to restart. This is useful when the server is restarting at
random times (say a hidden file is being created or modified in the
background). It also could allow you to check to see if you set up the
proper watch and ignore directories.